+++ R/iq-fast.R
@@ -8,7 +8,7 @@
 #
 # Pham TV, Henneman AA, Jimenez CR. iq: an R package to estimate relative protein abundances from ion quantification in DIA-MS-based proteomics, Bioinformatics 2020 Apr 15;36(8):2611-2613.
 #
-# Software version: 1.9.3
+# Software version: 1.9.6
 #
 #########################################################################
 
@@ -19,7 +19,10 @@ fast_read <- function(filename,
                       intensity_col = "F.PeakArea",
                       annotation_col = c("PG.Genes", "PG.ProteinNames"),
                       filter_string_equal = c("F.ExcludedFromQuantification" = "False"),
-                      filter_double_less = c("PG.Qvalue" = "0.01", "EG.Qvalue" = "0.01")) {
+                      filter_double_less = c("PG.Qvalue" = "0.01", "EG.Qvalue" = "0.01"),
+                      intensity_col_sep = NULL,
+                      intensity_col_id = NULL,
+                      na_string = "0") {
 
     cmd <- paste0("--sample ", sample_id,
                   " --primary ", primary_id,
@@ -42,10 +45,36 @@ fast_read <- function(filename,
         }
     }
 
+    if (!is.null(intensity_col_sep)) {
+        cmd <- paste0(cmd, " --intensity_col_sep ", intensity_col_sep)
+        cmd <- paste0(cmd, " --na_string ", na_string)
+    }
+
+    if (!is.null(intensity_col_id)) {
+        cmd <- paste0(cmd, " --intensity_col_id ", intensity_col_id)
+    }
+
     return(.Call("iq_filter", as.character(paste0(cmd, " ", filename))))
 }
 
 fast_MaxLFQ <- function(norm_data, row_names = NULL, col_names = NULL) {
+
+    # check for NA values
+    if (any(is.na(norm_data$protein_list))) {
+        stop("NA value in $protein_list.\n")
+    }
+
+    if (any(is.na(norm_data$sample_list))) {
+        stop("NA value in $sample_list.\n")
+    }
+
+    if (any(is.na(norm_data$id))) {
+        stop("NA value in $id.\n")
+    }
+
+    if (any(is.na(norm_data$quant))) {
+        stop("NA value in $quant.\n")
+    }
 
     if (is.null(row_names)) {
         proteins <- unique(as.character(norm_data$protein_list))
@@ -166,4 +195,79 @@ fast_preprocess <- function(quant_table,
     }
 
     return(d)
+}
+
+process_long_format <- function(input_filename,
+                                output_filename,
+                                sample_id = "File.Name",
+                                primary_id = "Protein.Group",
+                                secondary_id = "Precursor.Id",
+                                intensity_col = "Fragment.Quant.Corrected",
+                                annotation_col = NULL,
+                                filter_string_equal = NULL,
+                                filter_double_less = c("Q.Value" = "0.01", "PG.Q.Value" = "0.01"),
+                                intensity_col_sep = ";",
+                                intensity_col_id = NULL,
+                                na_string = "0",
+                                normalization = "median",
+                                log2_intensity_cutoff = 0,
+                                pdf_out = "qc-plots.pdf",
+                                pdf_width = 12,
+                                pdf_height = 8) {
+
+
+    if (normalization == "median") {
+        median_normalization <- TRUE
+    } else if (normalization == "none") {
+        median_normalization <- FALSE
+    } else {
+        stop("Unknown normalization method.")
+    }
+
+    iq_dat <- fast_read(input_filename,
+                        primary_id = primary_id,
+                        sample_id  = sample_id,
+                        secondary_id = secondary_id,
+                        intensity_col = intensity_col,
+                        intensity_col_sep = intensity_col_sep,
+                        annotation_col = annotation_col,
+                        filter_string_equal = filter_string_equal,
+                        filter_double_less = filter_double_less)
+
+    if (!is.null(iq_dat)) {
+        iq_norm_data <- fast_preprocess(iq_dat$quant_table,
+                                        median_normalization = median_normalization,
+                                        log2_intensity_cutoff = log2_intensity_cutoff,
+                                        pdf_out = pdf_out,
+                                        pdf_width = pdf_width,
+                                        pdf_height = pdf_height)
+
+        res <- fast_MaxLFQ(iq_norm_data,
+                           row_names = iq_dat$protein[, 1],
+                           col_names = iq_dat$sample)
+
+        message("Writing to: ", output_filename)
+
+        if (is.null(annotation_col)) {
+
+            tab <- cbind(rownames(res$estimate), res$estimate)
+            colnames(tab)[1] <- primary_id
+
+        } else {
+
+            extra_annotation <- extract_annotation(rownames(res$estimate),
+                                                   iq_dat$protein,
+                                                   primary_id = primary_id,
+                                                   annotation_columns = annotation_col)
+
+            tab <- cbind(rownames(res$estimate),
+                         extra_annotation[, annotation_col],
+                         res$estimate)
+            colnames(tab)[1:(length(annotation_col)+1)] <- c(primary_id, annotation_col)
+        }
+
+        write.table(tab, output_filename, sep = "\t", row.names = FALSE, quote = FALSE)
+    }
+
+    invisible(NULL)
 }