+++ R/iq.R
@@ -8,7 +8,7 @@
 #
 # Pham TV, Henneman AA, Jimenez CR. iq: an R package to estimate relative protein abundances from ion quantification in DIA-MS-based proteomics, Bioinformatics 2020 Apr 15;36(8):2611-2613.
 #
-# Software version: 1.9.3
+# Software version: 1.9.6
 #
 #########################################################################
 
@@ -21,29 +21,74 @@ preprocess <- function(quant_table,
                        log2_intensity_cutoff = 0,
                        pdf_out = "qc-plots.pdf",
                        pdf_width = 12,
-                       pdf_height = 8) {
+                       pdf_height = 8,
+                       intensity_col_sep = NULL,
+                       intensity_col_id = NULL,
+                       na_string = "0") {
 
     if (!is.null(pdf_out)) {
         pdf(pdf_out, pdf_width, pdf_height)
     }
 
-    if (class(quant_table[, intensity_col]) != "numeric") {
-        stop("Intensity column must be numeric.")
-    }
+    if (is.null(intensity_col_sep)) {
+        if (!is.numeric(quant_table[, intensity_col])) {
+            stop("Intensity column must be numeric when 'intensity_col_sep' is NULL.")
+        }
 
-    # make secondary ids
-    message("Concatenating secondary ids...\n")
-    second_id <- as.character(quant_table[, secondary_id[1]])
-    if (length(secondary_id) > 1) {
-        for (i in 2:length(secondary_id)) {
-            second_id <- paste(second_id, as.character(quant_table[, secondary_id[i]]), sep = "_")
+        # make secondary ids
+        message("Concatenating secondary ids...\n")
+        second_id <- as.character(quant_table[, secondary_id[1]])
+        if (length(secondary_id) > 1) {
+            for (i in 2:length(secondary_id)) {
+                second_id <- paste(second_id, as.character(quant_table[, secondary_id[i]]), sep = "_")
+            }
+        }
+
+        d <- data.frame(protein_list = as.character(quant_table[, primary_id]),
+                        sample_list = as.character(quant_table[, sample_id]),
+                        quant = log2(quant_table[, intensity_col]),
+                        id = second_id)
+    } else {
+        # make secondary ids
+        message("Concatenating secondary ids...\n")
+        second_id <- as.character(quant_table[, secondary_id[1]])
+        if (length(secondary_id) > 1) {
+            for (i in 2:length(secondary_id)) {
+                second_id <- paste(second_id, as.character(quant_table[, secondary_id[i]]), sep = "_")
+            }
+        }
+
+        v <- strsplit(as.character(quant_table[, intensity_col]), intensity_col_sep)
+        if (!is.null(intensity_col_id)) {
+            v_ids <- strsplit(as.character(quant_table[, intensity_col_id]), intensity_col_sep)
+        }
+
+        v_lengths <- lengths(v)
+        n_row <- sum(v_lengths)
+
+        d <- data.frame(protein_list = rep("", n_row),
+                        sample_list = rep("", n_row),
+                        quant = rep(0, n_row),
+                        id = rep("", n_row))
+        cc <- 1
+        for (r in 1:nrow(quant_table)) {
+            a <- v[[r]]
+            a[a == na_string] <- NA
+
+            cc_end <- cc + length(a) - 1
+
+            d$protein_list[cc:cc_end] <- as.character(quant_table[r, primary_id])
+            d$sample_list[cc:cc_end] <- as.character(quant_table[r, sample_id])
+            d$quant[cc:cc_end] <- log2(as.numeric(a))
+
+            if (is.null(intensity_col_id)) {
+                d$id[cc:cc_end] <- paste0(second_id[r], "_", 1:length(a))
+            } else {
+                d$id[cc:cc_end] <- paste0(second_id[r], "_", v_ids[[r]])
+            }
+            cc <- cc_end + 1
         }
     }
-
-    d <- data.frame(protein_list = as.character(quant_table[, primary_id]),
-                    sample_list = as.character(quant_table[, sample_id]),
-                    quant = log2(quant_table[, intensity_col]),
-                    id = second_id)
 
     d <- d[complete.cases(d),] # remove NaN quant
 
@@ -122,6 +167,23 @@ preprocess <- function(quant_table,
 
 create_protein_list <- function(preprocessed_data) {
 
+    # check for NA values
+    if (any(is.na(preprocessed_data$protein_list))) {
+        stop("NA value in $protein_list.\n")
+    }
+
+    if (any(is.na(preprocessed_data$sample_list))) {
+        stop("NA value in $sample_list.\n")
+    }
+
+    if (any(is.na(preprocessed_data$id))) {
+        stop("NA value in $id.\n")
+    }
+
+    if (any(is.na(preprocessed_data$quant))) {
+        stop("NA value in $quant.\n")
+    }
+
     proteins <- unique(preprocessed_data$protein_list)
     samples <- unique(preprocessed_data$sample_list)
 
@@ -154,7 +216,7 @@ create_protein_list <- function(preprocessed_data) {
                     m[char_r[j], char_c[j]] <- tmp[j, "quant"]
                 } else {
                     message(j, " : sample ", char_c[j], "; id ", char_r[j], " not unique.\n")
-                    stop("Error: duplicate entry.\n")
+                    stop("duplicate entry.\n")
                 }
             }
             p_list[[as.character(proteins[i])]] <- m
@@ -260,7 +322,7 @@ maxLFQ <- function(X) {
 
         res <- lsfit(rbind(cbind(2*AtA , rep(1, N)),
                            c(rep(1, N), 0)),
-                     c(2*Atb, mean(X, na.rm = TRUE) * N),
+                     c(2*Atb, mean(data.matrix(X), na.rm = TRUE) * N),
                      intercept = FALSE)
 
         return(res$coefficients[1:N])
@@ -377,7 +439,7 @@ create_protein_table <- function(protein_list, method = "maxLFQ", ...) {
         } else if (method == "topN") {
             out <- topN(protein_list[[i]], ...)
         } else if (method == "meanInt") {
-            out <- meanInt(protein_list[[i]])
+            out <- meanInt(protein_list[[i]], ...)
         } else {
             stop(paste0("Unknown method: ", method))
         }
@@ -409,4 +471,158 @@ extract_annotation <- function(protein_ids,
     rownames(tab) <- protein_ids
 
     return(tab)
+}
+
+process_wide_format <- function(input_filename,
+                                output_filename,
+                                id_column,
+                                quant_columns,
+                                data_in_log_space = FALSE,
+                                annotation_columns = NULL,
+                                method = "maxLFQ") {
+
+    message("Reading file: ", input_filename)
+    d <- read.delim(input_filename)
+    message("# rows = ", nrow(d), "; # cols = ", ncol(d))
+
+    if (!(id_column %in% colnames(d))) {
+        stop("'id_column' not found")
+    }
+
+    if (is.numeric(quant_columns)) {
+        if (max(quant_columns) > ncol(d)) {
+            stop("'quant_columns' out of range")
+        }
+        if (min(quant_columns) < 1) {
+            stop("'quant_columns' out of range")
+        }
+    } else {
+        if (!all(quant_columns %in% colnames(d))) {
+            stop("One or more column name in 'quant_columns' not found")
+        }
+    }
+
+    if (!all(annotation_columns %in% colnames(d))) {
+        stop("One or more column name in 'annotation_columns' not found")
+    }
+
+    message("Preparing data...")
+
+    # collapse to unique 'id_column'
+    d <- d[!is.na(d[, id_column]),]
+
+    d <- d[d[, id_column] != "",] # bad for rownames
+
+    d_long <- reshape(d,
+                      direction = "long",
+                      varying = quant_columns,
+                      timevar = "sample_list",
+                      times = colnames(d)[quant_columns],
+                      v.names = "quant")
+
+    d_long <- d_long[!is.na(d_long$quant) & !is.na(d_long[, id_column]),]
+
+    unique_values <- unique(d_long[, id_column])
+
+    if (!is.function(method) && method == "maxLFQ") {
+
+        df <- data.frame(protein_list = d_long[, id_column],
+                         sample_list = d_long$sample_list,
+                         id = d_long$id,
+                         quant = if (!data_in_log_space) log2(d_long$quant) else d_long$quant)
+
+        message("Running MaxLFQ...")
+
+        res <- fast_MaxLFQ(df)
+
+        res$estimate <- res$estimate[unique_values, ]
+
+    } else {
+
+
+        # set up result tables
+        res <- list(estimate = matrix(NA, nrow = length(unique_values), ncol = length(quant_columns)))
+
+        if (is.numeric(quant_columns)) {
+            colnames(res$estimate) <- colnames(d)[quant_columns]
+        }
+        else {
+            colnames(res$estimate) <- quant_columns
+        }
+        rownames(res$estimate) <- unique_values
+
+        thres_display <- length(unique_values)/20
+
+        message("Processing...")
+        for (r in 1:length(unique_values)) {
+
+            if (r > thres_display) {
+                message(format(r * 100/length(unique_values), digits = 2), "%")
+                thres_display <- r + length(unique_values)/20
+            }
+
+            a <- d[d[, id_column] == unique_values[r], quant_columns]
+
+            if (!data_in_log_space) {
+                a <- log2(a)
+            }
+
+            if (is.function(method)) {
+                out <- list(estimate = method(a))
+            } else if (method == "maxLFQ_R") {
+                out <- maxLFQ(a)
+            }
+            else if (method == "median_polish") {
+                out <- median_polish(a)
+            }
+            else if (method == "top3") {
+                out <- topN(a, N = 3)
+            }
+            else if (method == "top5") {
+                out <- topN(a, N = 5)
+            }
+            else if (method == "maxInt") {
+                out <- topN(a, N = 1)
+            }
+            else if (method == "meanInt") {
+                out <- meanInt(a)
+            }
+            else if (method == "sum") {
+                out <- list(estimate = log2(colSums(2^a, na.rm = TRUE)))
+            }
+            else if (method == "least_na") {
+                out <- list(estimate = a[which.min(rowSums(is.na(a))),])
+            }
+            else {
+                stop(paste0("Unknown method: ", method))
+            }
+
+            res$estimate[r,] <- as.numeric(out$estimate)
+        }
+    }
+
+    message("Writing to: ", output_filename)
+
+    if (is.null(annotation_columns)) {
+
+        tab <- cbind(rownames(res$estimate), res$estimate)
+        colnames(tab)[1] <- id_column
+
+    } else {
+
+        extra_annotation <- extract_annotation(rownames(res$estimate),
+                                               d,
+                                               primary_id = id_column,
+                                               annotation_columns = annotation_columns)
+
+        tab <- cbind(rownames(res$estimate),
+                     extra_annotation[, annotation_columns],
+                     res$estimate)
+        colnames(tab)[1:(length(annotation_columns)+1)] <- c(id_column, annotation_columns)
+
+    }
+
+    write.table(tab, output_filename, sep = "\t", row.names = FALSE, quote = FALSE)
+
+    invisible(NULL)
 }